+++ Cargo.toml
@@ -17,8 +17,6 @@ bench = false
 
 [lib]
 bench = false
-name = "nickel_lang"
-crate-type = ["cdylib", "lib"]
 
 [features]
 default = ["markdown", "repl", "doc"]
@@ -26,7 +24,6 @@ markdown = ["termimad"]
 repl = ["rustyline", "rustyline-derive", "ansi_term"]
 repl-wasm = ["wasm-bindgen", "js-sys", "serde_repr"]
 doc = [ "comrak" ]
-python-module = [ "pyo3" ]
 
 [build-dependencies]
 lalrpop = "0.19.6"
@@ -67,8 +64,6 @@ comrak = { version = "0.12.1", optional = true, features = [] }
 once_cell = "1.14.0"
 typed-arena = "2.0.1"
 
-pyo3 = { version = "0.17.3", optional = true, features = ["extension-module"] }
-
 [dev-dependencies]
 pretty_assertions = "1.2.1"
 assert_matches = "1.5.0"
@@ -85,6 +80,7 @@ members = [
     "lsp/nls",
     "utilities",
     "nickel-wasm-repl",
+    "pyckel",
 ]
 
 # Enable this to use flamegraphs
