+++ pyckel/pyproject.toml
@@ -3,5 +3,4 @@ requires = ["maturin>=0.14,<0.15"]
 build-backend = "maturin"
 
 [tool.maturin]
-features = ["python-module"]
