+++ src/lib.rs
@@ -10,8 +10,6 @@ pub mod parser;
 pub mod position;
 pub mod pretty;
 pub mod program;
-#[cfg(feature = "python-module")]
-pub mod python;
 pub mod repl;
 pub mod serialize;
 pub mod stdlib;
